feat(records): add submenus and default booking records
