+++ src/components/RoomBooking/Records/AccessClassroomRecords.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="placeholder">
+    教室出入记录内容
+  </div>
+</template>
+
+<script setup>
+</script>
+
+<style scoped>
+.placeholder {
+  padding: 20px;
+}
+</style>
