+++ src/components/RoomBooking/Records/AccessRemoteRecords.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="placeholder">
+    远程开门记录内容
+  </div>
+</template>
+
+<script setup>
+</script>
+
+<style scoped>
+.placeholder {
+  padding: 20px;
+}
+</style>
